Updated main.py
Updates include:
(1) only run OCR on files that have not been analyzed yet 
(2) take a mode argument into ocr_catalog that, if set mode="ocr_no" then we do not run OCR, we skip that step and go directly to clustering already-extant local text files
(3) writes the WordSpace data to a file and reconstructs it from file as well, so that after adding books to a wordspace the first time, we can reconstruct that a bit more quickly
(4) Added some output lines so we can monitor clustering a bit more closely.

Problems:
If you have a searchable PDF, it already has all the text. Currently, we run OCR on it anyway and store that text. This is a waste of computational resources, but also has the strange effect of doubling the text in the document. Since we are only comparing histograms of words, this has no net effect on our clustering, but it's a waste of resources and it's a source of data corruption. Ideally we would like a way to identify whether a PDF is already in text form before performing OCR...